extract util logging
